added railinsights as project and renamed SingleProject component to ProjectCard
