+++ src/lib/ProjectCard.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import type { Project } from "../lib/types";
+  import type { Project } from "./types";
   export let project: Project;
 </script>
