+++ src/lib/projects.ts
@@ -2,6 +2,16 @@ import type { Project } from "../lib/types";
 
 export const projects: Project[] = [
   {
+    heading: "Rail Insights",
+    descriptions: [
+      "Rail Insights is a web application that focuses on the evaluation and statistical representation of Deutsche Bahn trains. The application allows users to gain insight into statistics on trains, stations, and global delays. In addition, the application offers a forecast function for future journeys.",
+      "The project uses SpringBoot, MariaDB and Vaadin for the frontend. 🚆",
+    ],
+    projectLink: "https://www.railinsights.de",
+    imageURL: "images/rli-favicon.png",
+    displayColor: "#F50909",
+  },
+  {
     heading: "MensiMates",
     descriptions: [
       "MensiMates is an application that helps you stay up-to-date with the meals in your campus cafeteria. Never miss a meal again, and enjoy a variety of options at your fingertips.",
